Enhanced Ganache deployment instructions.
The tutorial did  not instruct the user on how to add a section to their Truffle configuration file that describes the Ganache network.  This version has those instructions.  Also instead of using "local" for the deployment command, this version uses "development" as the network mnemonic since that is the default label used in Truffle's documentation on the Truffle configuration file.